Disable switching to build-agent user on launch
This commit disables automatically switching to the `build-agent`,
a non-root user, context on launch because doing so can corrupt the
permissions of the files created inside the GitHub Actions runner.

This should be kept disabled until a fundamental solution to this
problem is provided by the GitHub, or a usable workaround is found.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>