Dockerfile: Add MinGW-w64 toolchain
This commit adds a script that builds MinGW-w64 toolchains with `win32`
and `posix` thread models and updates the Dockerfile to invoke it.

The MinGW-w64 toolchains are installed in the Docker image as follows:

* /opt/mingw-w64-win32: x86_64-w64-mingw32 toolchain, win32 thread model
* /opt/mingw-w64-posix: x86_64-w64-mingw32 toolchain, posix thread model

The toolchain build scripts cleans up after itself so that no
intermediate build artifacts end up in the final Docker image.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>